all in one, rough version
